Added support for latest node and android 28 sdk
We changed the way to load a non-cordova module q. This was breaking when we were providing support for targetsdk v28. 

Following error was reported in console for the version we were using (v2.4.0).
Failed to install 'cordova-sqlite-storage': CordovaError: Using "requireCordovaModule" to load non-cordova module "q" is not supported. Instead, add this module to your dependencies and use regular "require" to load it